- get simulations (client side) : fixed - display customer's simulations : moving some operations to the controller - new model reset options
